Files considering environment are now much better organized (following object-oriented programming). Other minor modifications are made consequently.

+++ include/state_spaces/Tree.h
@@ -20,27 +20,27 @@ namespace base
 	{
 	private:
 		std::string tree_name;
-		uint tree_idx;
+		int tree_idx;
 		std::shared_ptr<std::vector<std::shared_ptr<base::State>>> states; 	// List of all nodes in the tree
         std::shared_ptr<base::KdTree> kd_tree;
 
 	public:
 		Tree() {}
-		Tree(const std::string &tree_name_, uint tree_idx_);
+		Tree(const std::string &tree_name_, int tree_idx_);
 		Tree(const std::shared_ptr<std::vector<std::shared_ptr<base::State>>> states_);
 		~Tree();
 
 		inline const std::string &getTreeName() const { return tree_name; }
-		inline const uint getTreeIdx() const { return tree_idx; }
+		inline const int getTreeIdx() const { return tree_idx; }
 		inline std::shared_ptr<std::vector<std::shared_ptr<base::State>>> getStates() const { return states; }
-		inline std::shared_ptr<base::State> getState(size_t idx) const { return states->at(idx); }
+		inline std::shared_ptr<base::State> getState(int idx) const { return states->at(idx); }
         inline std::shared_ptr<base::KdTree> getKdTree() const { return kd_tree; }
-		inline size_t getNumStates() const { return states->size(); }
+		inline int getNumStates() const { return states->size(); }
 
 		inline void setTreeName(const std::string &tree_name_) { tree_name = tree_name_; }
-		inline void setTreeIdx(const uint treeIdx_) { tree_idx = treeIdx_; }
+		inline void setTreeIdx(const int treeIdx_) { tree_idx = treeIdx_; }
 		inline void setStates(const std::shared_ptr<std::vector<std::shared_ptr<base::State>>> states_) { states = states_; }
-		inline void setState(const std::shared_ptr<base::State> state, size_t idx) { states->at(idx) = state; }
+		inline void setState(const std::shared_ptr<base::State> state, int idx) { states->at(idx) = state; }
         inline void setKdTree(const std::shared_ptr<base::KdTree> kdtree_) { kd_tree = kdtree_; }
 
 		void clearTree();
@@ -52,8 +52,8 @@ namespace base
 
 		template <class BBOX> 
         bool kdtree_get_bbox(BBOX& /* bb */) const { return false; }
-		inline size_t kdtree_get_point_count() const { return states->size(); }
-		inline float kdtree_get_pt(const size_t idx, const size_t dim) const { return states->at(idx)->getCoord()[dim]; }
+		inline int kdtree_get_point_count() const { return states->size(); }
+		inline float kdtree_get_pt(const int idx, const int dim) const { return states->at(idx)->getCoord()[dim]; }
 
 		friend std::ostream &operator<<(std::ostream &os, const Tree &tree);
 	};
